refactor: reuse sequence dataset helper in inference
